refactor: re-layout of timeline and scss global refactor
