+++ src/scss/Projects.scss
@@ -1,75 +1,34 @@
 @import 'colors.scss';
 
-.projects-tag.application,
-.projects-tag.game,
-.projects-tag.web,
-.projects-tag.ip,
-.projects-tag.library,
-.projects-tag.simulator {
+.tag.application,
+.tag.game,
+.tag.web,
+.tag.ip,
+.tag.library,
+.tag.simulator {
   background-color: $industry-knowledge-color;
 }
 
-.projects-tag.computer-vision,
-.projects-tag.machine-learning,
-.projects-tag.side-channel {
+.tag.computer-vision,
+.tag.machine-learning,
+.tag.side-channel {
   background-color: $scientific-knowledge-color;
 }
 
-.projects-tag.python,
-.projects-tag.rust,
-.projects-tag.typescript,
-.projects-tag.c\+\+,
-.projects-tag.c,
-.projects-tag.javascript {
+.tag.python,
+.tag.rust,
+.tag.typescript,
+.tag.c\+\+,
+.tag.c,
+.tag.javascript {
   background-color: $software-languages-color;
 }
 
-.projects-tag.numpy,
-.projects-tag.scipy,
-.projects-tag.piston,
-.projects-tag.react,
-.projects-tag.npm,
-.projects-tag.fpga {
+.tag.numpy,
+.tag.scipy,
+.tag.piston,
+.tag.react,
+.tag.npm,
+.tag.fpga {
   background-color: $tools-technologies-color;
-}
-
-.projects-list {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-}
-
-.projects-tags-list {
-  display: flex;
-  align-content: center;
-}
-
-.projects-tag {
-  margin: 0 4px 0 0;
-  font-size: 12px;
-  font-weight: bold;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  color: white;
-  padding: 4px;
-  background-color: $theme-gray-700;
-}
-
-.projects-item {
-  flex-basis: 30%;
-  margin: 8px;
-  padding: 16px;
-  min-width: 300px;
-  background-color: $theme-gray-300;
-}
-
-.projects-text {
-  font-weight: normal;
-  font-style: italic;
-}
-
-@media only screen and (max-width: 1024px) {
-  .projects-item {
-    flex-grow: 1;
-  }
 }
