+++ src/scss/Skills.scss
@@ -97,7 +97,7 @@
   flex-direction: column;
 }
 
-@media only screen and (max-width: 767px) {
+@media only screen and (max-width: 1024px) {
   #skills-inner {
     flex-direction: column;
     align-items: start;
