+++ src/scss/Timeline.scss
@@ -1,200 +1,9 @@
 @import 'colors.scss';
 
-$barColor: $theme-gray-300;
-$timelineBackground: $theme-gray-300;
-
-.timeline-item {
-  h4 {
-    margin-top: 4px;
-  }
-}
-
-.collapse-text {
-  p, ul {
-  text-align: left;
-  vertical-align: bottom;
-  margin-top: 0;
-  margin-bottom: 0;
-  }
-
-  ul {
-    padding-left: 32px;
-  }
-}
-
-.timeline-container {
-  display: flex;
-  flex-direction: column;
-  position: relative;
-}
-
-.timeline-container::after {
-  background-color: $barColor;
-  content: '';
-  position: absolute;
-  left: calc(50% - 2px);
-  width: 4px;
-  height: 100%;
-}
-
-.timeline-item {
-  display: flex;
-  justify-content: flex-end;
-  padding-right: 30px;
-  position: relative;
-  margin: 10px 0;
-}
-
-
-.timeline-item-content {
-  border-radius: 10px;
-  background-color: $timelineBackground;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  padding: 18px;
-  position: relative;
-  width: 100%;
-  text-align: right;
-}
-
-.timeline-item-content::after {
-  content: ' ';
-  background-color: $timelineBackground;
-  position: absolute;
-  right: -7.5px;
-  top: calc(50% - 7.5px);
-  transform: rotate(45deg);
-  width: 15px;
-  height: 15px;
-}
-
-
-
-.timeline-tag {
-  font-size: 12pt;
-}
-
-.timeline-item-content .timeline-tag {
-  color: white;
-  font-weight: bold;
-  top: 5px;
-  left: 5px;
-  letter-spacing: 1px;
-  padding: 4px;
-  position: absolute;
-  text-transform: uppercase;
-  border-radius: 10px;
-}
-
-
-
-.timeline-item-content time {
-  color: $theme-gray-700;
-  font-size: 12pt;
-  font-weight: bold;
-}
-
-.timeline-item-content p, .timeline-item-content ul {
-  line-height: 24px;
-}
-
-.timeline-item-content .circle {
-  background-color: white;
-  border: 3px solid $barColor;
-  border-radius: 50%;
-  position: absolute;
-  top: calc(50% - 10px);
-  right: -40px;
-  width: 20px;
-  height: 20px;
-  z-index: 100;
-}
-
-@media only screen and (min-width: 768px) {
-  .timeline-item:nth-child(odd) .timeline-item-content {
-    text-align: left;
-    align-items: flex-start;
-  }
-
-  .timeline-item:nth-child(odd) .timeline-item-content::after {
-    right: auto;
-    left: -7.5px;
-  }
-
-  .timeline-item:nth-child(odd) .timeline-item-content .timeline-tag {
-    left: auto;
-    right: 5px;
-  }
-
-  .timeline-item:nth-child(odd) .timeline-item-content .circle {
-    right: auto;
-    left: -40px;
-  }
-
-  .timeline-item:nth-child(odd) {
-    align-self: flex-end;
-    justify-content: flex-start;
-    padding-left: 30px;
-    padding-right: 0;
-  }
-
-
-}
-
-@media only screen and (max-width: 1023px) {
-  .timeline-item-content {
-    max-width: 100%;
-  }
-}
-
-@media only screen and (max-width: 767px) {
-  .timeline-item-content,
-  .timeline-item:nth-child(odd) .timeline-item-content {
-    padding: 15px 10px;
-    text-align: center;
-    align-items: center;
-
-    .timeline-item {
-      width: 100%;
-    }
-  }
-
-  .timeline-item-content .timeline-tag {
-    text-align: center;
-  }
-
-  .timeline-item-content time {
-    margin-top: 20px;
-  }
-
-  .timeline-container::after {
-    left: 100%;
-  }
-}
-
-@media only screen and (min-width: 768px) {
-  .timeline-container::after {
-    left: calc(50% - 2px);
-  }
-
-  .timeline-item {
-    width: 50%;
-  }
-}
-
-.experience {
+.tag.experience {
   background-color: $theme-tag-red;
 }
 
-.education {
+.tag.education {
   background-color: $theme-tag-purple;
-}
-
-.timeline-item-brief {
-  font-style: italic;
-}
-
-#timeline p {
-  margin-bottom: 2px;
 }
