+++ src/timeline/TimelineList.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import TimelineItem from './TimelineItem'
+import '../scss/Backlog.scss'
 
 const TimelineList = ({items}) => {
     const sortedItems = items.slice().sort((a, b) => new Date(b.start) - new Date(a.start));
-    return (<div className="timeline-container">
+    return (<div className="backlog-list">
         {sortedItems.map((item) => (
             <TimelineItem
                 category={item.category}
@@ -12,8 +13,8 @@ const TimelineList = ({items}) => {
                 items={item.items}
                 start={item.start}
                 end={item.end}
-                key={item.title + "&" + item.start}
-                id={item.title + "&" + item.start}
+                key={item.title + item.start}
+                id={item.title + item.start}
                 company={item.company}
                 location={item.location}
             />
